Create 14.Longest_Common_Prefix.py
Not that good of a solution but it works.

There were better functions like startswith - You could set the first term as a prefix and compare it will all other terms. If no match, then cut the last character from prefix till it matches. If it becomes empty, return that. Or return the prefix after the loop process is completed.